pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ark"
version = "0.1.0"
description = "Ark backup agent for AI-assisted system backup"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Ark Contributors" }
]
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.1",
  "questionary>=2.1.0",
  "pathspec>=0.12.1",
  "tomli>=2.0.1; python_version < '3.11'",
  "litellm>=1.74.0",
  "pydantic>=2.8.0",
  "google-auth>=2.39.0",
  "google-auth-oauthlib>=1.2.2",
  "google-genai>=0.8.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-mock>=3.14.0"
]

[project.scripts]
ark = "src.cli:main"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
